chore(application-estate): Simplify service and standardize upload data (#12460)
* WIP - splitting up code for better maintainability

* Finalize expanders and upload data pipeline

* Add estateWithoutAsset fields and filtering empty objects

* Remove unused code

* Fix some filtering issues with empty default objects

* chore: nx format:write update dirty files

* Fix linting error

---------
